--- src/main/java/com/etraveli/practice/consumer/GenericConsumer.java
@@ -1,7 +1,0 @@
-package com.etraveli.practice.consumer;
-
-import java.util.function.Consumer;
-
-public interface GenericConsumer<T> extends Consumer<T> {
-
-}
