+++ src/main/java/com/etraveli/practice/service/impl/RentalInfoServiceImpl.java
@@ -11,6 +11,7 @@ import com.etraveli.practice.util.CustomerUtils;
 import com.etraveli.practice.util.MovieRentalUtils;
 import com.etraveli.practice.service.RentalInfoService;
 import com.google.common.util.concurrent.AtomicDouble;
+import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -36,22 +37,25 @@ public class RentalInfoServiceImpl implements RentalInfoService {
         return buildOutput(customer, customerRentalSummary);
     }
 
+    /*P.S. The use of vars ("var name = new ...") over explicit types is down to personal preference and project guidelines. I am demonstrating it here for the sake of readability and conciseness.
+    And since we already have Lombok, we can even use "val" to make it equivalent to "final var"! In all honesty, I should have been declaring more of the fields as final, but was a bit lazy...*/
     public static CustomerRentalSummary processMovieRentalRecords(Customer customer) {
-        AtomicDouble totalDebtAmountOwed = new AtomicDouble(0);
-        AtomicInteger totalFrequentPoints = new AtomicInteger();
-        List<MovieRentalDebtRecord> movieRentalDebtRecords = Collections.synchronizedList(new LinkedList<>());
+        val totalDebtAmountOwed = new AtomicDouble(0);
+        val totalFrequentPoints = new AtomicInteger(0);
+        val movieRentalDebtRecords = Collections.synchronizedList(new LinkedList<MovieRentalDebtRecord>());
 
         customer.movieRentals()
                 .parallelStream()
                 .forEach(movieRental -> populateMovieRentalDebtRecords(movieRental, totalDebtAmountOwed, totalFrequentPoints, movieRentalDebtRecords));
 
+        // Sort the records by debt owed in descending order, ensuring that titles with the most debt are listed first
         movieRentalDebtRecords.sort(Comparator.comparing(MovieRentalDebtRecord::debtOwedForTitle).reversed());
         return new CustomerRentalSummary(totalDebtAmountOwed, totalFrequentPoints, movieRentalDebtRecords);
     }
 
     public static void populateMovieRentalDebtRecords(MovieRental movieRental, AtomicDouble totalDebtAmountOwed, AtomicInteger totalFrequentPoints, List<MovieRentalDebtRecord> movieRentalDebtRecords) {
-        Double debtOwedForTitle = MovieRentalUtils.calculateDebtOwedForMovieRental(movieRental);
-        totalDebtAmountOwed.updateAndGet(v -> v + debtOwedForTitle);
+        val debtOwedForTitle = MovieRentalUtils.calculateDebtOwedForMovieRental(movieRental);
+        totalDebtAmountOwed.getAndAdd(debtOwedForTitle);
         totalFrequentPoints.getAndAdd(MovieRentalUtils.calculateFrequentPoints(movieRental));
         movieRentalDebtRecords.add(new MovieRentalDebtRecord(MovieEnum.valueOf(movieRental.movieId()).getTitle(), debtOwedForTitle));
     }
