--- src/main/java/com/etraveli/practice/supplier/GenericSupplier.java
@@ -1,7 +1,0 @@
-package com.etraveli.practice.supplier;
-
-import java.util.function.Supplier;
-
-public interface GenericSupplier<T> extends Supplier<T> {
-
-}
